Separate DataTableHeader into separate component
